:hammer: Extract Cauldron types
