+++ ern-cauldron-api/src/types/Cauldron.ts
@@ -1,0 +1,6 @@
+import { CauldronNativeApp } from './CauldronNativeApp'
+export interface Cauldron {
+  schemaVersion: string
+  config?: any
+  nativeApps: CauldronNativeApp[]
+}
