+++ ern-cauldron-api/src/types/CauldronCodePushEntry.ts
@@ -1,0 +1,6 @@
+import { CauldronCodePushMetadata } from './CauldronCodePushMetadata'
+export interface CauldronCodePushEntry {
+  metadata: CauldronCodePushMetadata
+  miniapps: string[]
+  jsApiImpls: string[]
+}
