+++ ern-cauldron-api/src/types/CauldronCodePushMetadata.ts
@@ -1,0 +1,12 @@
+export interface CauldronCodePushMetadata {
+  deploymentName: string
+  isMandatory?: boolean
+  isDisabled?: boolean
+  appVersion?: string
+  size?: number
+  releaseMethod?: string
+  label?: string
+  releasedBy?: string
+  rollout?: number
+  promotedFromLabel?: string
+}
