+++ ern-cauldron-api/src/types/CauldronNativeApp.ts
@@ -1,0 +1,7 @@
+import { CauldronNativeAppPlatform } from './CauldronNativeAppPlatform'
+
+export interface CauldronNativeApp {
+  name: string
+  platforms: CauldronNativeAppPlatform[]
+  config?: any
+}
