+++ ern-cauldron-api/src/types/CauldronNativeAppPlatform.ts
@@ -1,0 +1,7 @@
+import { CauldronNativeAppVersion } from './CauldronNativeAppVersion'
+
+export interface CauldronNativeAppPlatform {
+  name: string
+  versions: CauldronNativeAppVersion[]
+  config?: any
+}
