+++ ern-cauldron-api/src/types/CauldronNativeAppVersion.ts
@@ -1,0 +1,12 @@
+import { CauldronContainer } from './CauldronContainer'
+
+export interface CauldronNativeAppVersion {
+  name: string
+  isReleased: boolean
+  binary?: string
+  yarnLocks: any
+  container: CauldronContainer
+  codePush: any
+  config?: any
+  containerVersion: string
+}
