+++ ern-cauldron-api/src/types/ICauldronDocumentAccess.ts
@@ -1,0 +1,6 @@
+import { Cauldron } from './Cauldron'
+
+export interface ICauldronDocumentAccess {
+  commit(message: string): Promise<void>
+  getCauldron(): Promise<Cauldron>
+}
