+++ ern-cauldron-api/src/types/ICauldronDocumentStore.ts
@@ -1,0 +1,4 @@
+import { ITransactional } from './ITransactional'
+import { ICauldronDocumentAccess } from './ICauldronDocumentAccess'
+
+export type ICauldronDocumentStore = ITransactional & ICauldronDocumentAccess
