+++ ern-cauldron-api/src/types/ICauldronFileStore.ts
@@ -1,0 +1,4 @@
+import { ITransactional } from './ITransactional'
+import { ICauldronFileAccess } from './ICauldronFileAccess'
+
+export type ICauldronFileStore = ITransactional & ICauldronFileAccess
