+++ ern-cauldron-api/src/types/ITransactional.ts
@@ -1,0 +1,5 @@
+export interface ITransactional {
+  beginTransaction(): Promise<void>
+  commitTransaction(message: string | string[]): Promise<void>
+  discardTransaction(): Promise<void>
+}
